% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MedMR.r
\name{MedMR}
\alias{MedMR}
\title{Multiply robust estimation of natural indirect effects with multiple ordered mediators.}
\usage{
MedMR(
  data,
  exposure,
  mediators,
  outcome,
  confounder,
  m_type,
  y_type = "discrete",
  method = "Robust",
  scale = "difference",
  boot = FALSE,
  boot_ratio = 0.8,
  boot_rep = 10,
  seed_num = 123,
  double_mont = 1e+05,
  single_mont = 2e+05,
  b_rep = TRUE
)
}
\arguments{
\item{data}{a data frame, where the column is the variable and row is the sample.}

\item{exposure}{variable name of the exposure.}

\item{mediators}{variable names of mediators.}

\item{outcome}{variable name of the outcome.}

\item{confounder}{variable names of confounders.}

\item{m_type}{a description of the mediator model. This should be one of "d" (logistic model), "normal" (Gaussian model), "exp" (exponential model), and "poisson" (poisson model).}

\item{y_type}{a description of the outcome model. "discrete" uses logistic model, and "continuous" uses Gaussian model.}

\item{method}{a description of the estimation method. "IPW" stands for the inverse probability weighting; "G" stands for G-computation; "Robust" stands for multiply robust estimation. Default is "Robust".}

\item{boot}{a logical value. if 'TRUE' nonparametric bootstrap will be used for SD estimation. Default is 'FALSE'.}

\item{boot_ratio}{a ratio of the random samples of individuals from the original data.}

\item{boot_rep}{a number of replicates}

\item{seed_num}{seed. Default is 123.}

\item{double_mont}{a number of Monte Carlo simulations in multiply robust estimation.}

\item{single_mont}{a number of Monte Carlo simulations in G-computation.}

\item{b_rep}{if true resampling with replacement will be used; if false resampling without replacement will be used.}
}
\value{
A list of estimated direct, indirect, and total effects.
TE is the total effect. Seq0 is the natural direct effect.
Seq1,...SeqK are the natural indirect effects of mediators.
}
\description{
MedMR is used to assess the natural indirect effects with multiple mediators by multiply robust estimation, G-computation, and inverse probability weighting.
}
\examples{
data("HCCexample")
ff <- MedMR(data=HCCexample,exposure="a",mediators=c("m1","m2"),outcome="y",
confounder=c("gender","age","smoke","alcohol"),
m_type=c("d","d"),y_type = "discrete",
method = "Robust",
scale = "difference",
boot = T,
boot_ratio = 0.8,
boot_rep = 20,
seed_num = 123,
double_mont = 1e5,
single_mont = 2e5,
b_rep = TRUE)
}
\author{
An-Shun Tai \email{anshuntai@nctu.edu.tw} Chen-Chien Lee\email{xu3t6fu0@gmail.com}
}
